refactor: 카드 전달 책임 이동
